refactor(runtime): extract web_legacy / meta / sandbox tool modules
Three more leaf domains out of tool_runner/mod.rs:

- `web_legacy` — tool_web_fetch_legacy + tool_web_search_legacy, the
  WebToolsContext-less fallbacks. parse_ddg_results is consumed here
  now, so the corresponding import on mod.rs goes away.
- `meta` — tool_load + tool_search + the meta_lookup_pool helper.
  Pulls super::builtin_tool_definitions() so an empty granted slice
  still reports "not found" and only a None slice falls back to the
  builtin catalog (#3044). ToolExecutionStatus moves with it.
- `sandbox` — tool_docker_exec. The Docker module itself stays in
  crate::docker_sandbox; this only carves out the dispatch-side wrapper.

Verbatim function moves; visibility upgraded to pub(super). No
behavior change.

mod.rs: 11,744 -> 11,449 LOC (-295). cargo check + cargo clippy --lib
both clean.

Refs #3710
